Move the sheet action row into the components package
